bfa: fix faulty handling of events in lps sm
Patchwork-id: 60514
O-Subject: [RHEL6.5 PATCH 25/43] scsi: bfa: fix faulty handling of events in lps sm
Bugzilla: 928983
RH-Acked-by: Rob Evers <revers@redhat.com>

Upstream Status:
b2b1a6aadaa11b48c7d87273f41fd7720e1f8a4d

When a switch disable/enable or a reboot is done, the HBA port gets an
offline and a subsequent online notification. When the port comes up a
link up notification is sent to bfa from the firmware. The bfa then send
an FLOGI to the firmware which is sent out on the wire.
The switch port meanwhile goes offline (presumably for diagnostics)
which causes the switch not to respond to the FLOGI.
The link down notification is sent to the HBA driver. However owing to a
bug in the lps state machine handling the lps state machine does not
move to sm_init state (it remains in sm_login state and send a login
complete message to fcs). This results in a zero PID assignment as the
login is not really complete.

This fix is to correctly handle the events in lps state machine.

Signed-off-by: Vijaya Mohan Guvva <vmohan@brocade.com>
Signed-off-by: James Bottomley <JBottomley@Parallels.com>
Signed-off-by: Rafael Aquini <aquini@redhat.com>