bfa: Fix 1860 port initialize when ATC is enabled
Patchwork-id: 60520
O-Subject: [RHEL6.5 PATCH 32/43] scsi: bfa: Fix 1860 port initialize when ATC is enabled
Bugzilla: 928983
RH-Acked-by: Rob Evers <revers@redhat.com>

Upstream Status:
http://marc.info/?l=linux-scsi&m=136843872227448&w=2

On Xen kernels, if ATC (address translation cache) is enabled, the
first PCIe DMA read from the adapter fails with an error. This is due to
a bug ASIC, which leads to a failure of 1860 ports to be initialised.

This patch includes the fix to disable Invalidated Tag Match Enable
capability by setting the bit 26 of CHIP_MISC_PRG to 0, by default it is
set to 1.

Signed-off-by: Vijaya Mohan Guvva <vguvva@redhat.com>
Signed-off-by: Rafael Aquini <aquini@redhat.com>