Revert: kernel: add support for physical memory > 4TB
Patchwork-id: 60540
O-Subject: [RHEL6 PATCH] Revert "[s390] kernel: add support for physical memory > 4TB"
Bugzilla: 913220

This reverts commit 394f1b0984c610ad2d39d72ca8db95ab645667ba.

Bugzilla: 913220

There are required kexec-tools userspace changes required to make kdump
functional with this change. They won't be ready for a bit longer, so in
the interim, we're going to revert this patch so that we can continue
testing s390 in beaker with working kdump support.

CC: Hendrik Brueckner <brueckner@redhat.com>
Signed-off-by: Jarod Wilson <jarod@redhat.com>