bnx2x: Change MDIO clock settings
Patchwork-id: 60543
O-Subject: [RHEL6 PATCH] bnx2x: Change MDIO clock settings
Bugzilla: 901747
RH-Acked-by: Ivan Vecera <ivecera@redhat.com>
RH-Acked-by: Veaceslav Falico <vfalico@redhat.com>
RH-Acked-by: Jes Sorensen <Jes.Sorensen@redhat.com>
RH-Acked-by: David S. Miller <davem@redhat.com>

From: Yaniv Rosner <yanivr@broadcom.com>

BZ: https://bugzilla.redhat.com/show_bug.cgi?id=901747

A customer is seeing occasional MDC/MDIO timeout errors reported by the bnx2x
driver accompanied by the loss of the link. Broadcom recognized it as a known
issue with contention on the MDC bus. The corresponding fix is upstream:

commit 55386fe88349706ae570522180d89892883af2b5
Author: Yaniv Rosner <yanivr@broadcom.com>
Date:   Tue Nov 27 03:46:30 2012 +0000

    bnx2x: Change MDIO clock settings

    When drivers works on top of an old bootcode, it is theoretically subjected to
    MDC/MDIO failures since the MDIO clock is set in the beginning of each sequence,
    rather than per CL45 command. On rare cases an old bootcodes may change that in
    the middle, so to address that, the MDIO clock is set for each CL45 access.
    In addition, setting the MDIO clock is now done per EMAC base, and
    not per port number, since a specific port can potentially use both EMACs for
    different PHY accesses.

    Signed-off-by: Yaniv Rosner <yanivr@broadcom.com>
    Signed-off-by: Eilon Greenstein <eilong@broadcom.com>
    Signed-off-by: David S. Miller <davem@davemloft.net>

Brew build:
https://brewweb.devel.redhat.com/taskinfo?taskID=5923954

Testing was performed by the customer. They were willing to test a
patched bnx2x.ko module on top of a RHEL 6.2 kernel that they were
using (Note: The patch had to be backported to RHEL 6.2 for this test
and it is attached in the BZ). They confirmed the problem could not be
reproduced with the patched module.

Signed-off-by: Jarod Wilson <jarod@redhat.com>