GFS2: Remove inode from i_ordered list in gfs2_clear_inode
Patchwork-id: 60609
O-Subject: [RHEL6.5 PATCH bz859444] GFS2: Remove inode from i_ordered list in gfs2_clear_inode
Bugzilla: 859444
RH-Acked-by: Abhijith Das <adas@redhat.com>
RH-Acked-by: Andrew Price <anprice@redhat.com>

Hi,

While testing GFS2, QE discovered a regression in patch #7 for
bug #859444, which can be seen here:

http://git.app.eng.bos.redhat.com/?p=rhel6-fs.git;a=commitdiff;h=71e2a08c3c17e97a3a667e9d22469a9c185b7a7c

The problem was introduced due to a difference between
upstream and RHEL6. In upstream, tearing down of inodes is
handled differently. Function clear_inode in vfs no longer
calls a fs-specific helper function like it does in RHEL6.

Since the RHEL6 patch was a strict 1:1 translation of the
upstream patch, it did not take into account that the helper
function, gfs2_clear_inode was called in RHEL6. Due to the
change in calling sequence of inode tear-down, we need to make
the helper function remove the inode from the ordered write list.
Since we didn't remove it from the list, a different inode
encountered it there and tried to sync its data. The trouble is,
by then, the inode had been freed, so it segfaulted when it
tried to reference it.

This patch introduces the necessary call to remove the inode
being freed from the ordered list, so it should not be found
by other inodes during the sync process.

This patch removes inodes from the ordered writes list in function
gfs2_clear_inode, which is called when vfs instructs GFS2 to tear
down a dinode because it's no longer needed. This is necessary to
prevent others from trying to access the inode after it's been
freed.

The problem only affects RHEL6, and the patch is not needed
upstream.

rhbz#859444

Regards,

Bob Peterson
Red Hat File Systems

Signed-off-by: Bob Peterson <rpeterso@redhat.com>
Signed-off-by: Jarod Wilson <jarod@redhat.com>