fix race in scsi_target_reap
Patchwork-id: 60634
O-Subject: [RHEL6.5 PATCH] scsi: fix race in scsi_target_reap
Bugzilla: 910203
RH-Acked-by: Ewan Milne <emilne@redhat.com>
RH-Acked-by: Jes Sorensen <Jes.Sorensen@redhat.com>
RH-Acked-by: Jan Vesely <jvesely@redhat.com>

This is backported from upstream
commit f9e8894ae5157796dd69249c56062042d02a431d
The reporter hasn't reported back if it solves his problem, but
the patch fixes a race in a way which seems safe, so it is useful anyway.

This patch (as1357) fixes a race in SCSI target allocation and
release.  Putting a target in the STARGET_DEL state isn't protected by
the host lock, so an old target structure could be reused by a new
device even though it's about to be deleted.  The cure is to change
the state while still holding the host lock.

Signed-off-by: Alan Stern <stern@rowland.harvard.edu>
Signed-off-by: James Bottomley <James.Bottomley@suse.de>
Signed-off-by: Rafael Aquini <aquini@redhat.com>