perf: Fix offcore_rsp valid mask for SNB/IVB
Patchwork-id: 60716
O-Subject: [PATCH RHEL6.5 BZ971315 CVE-2013-2146] perf/x86: Fix offcore_rsp valid mask for SNB/IVB
Bugzilla: 971315
CVE: CVE-2013-2146
RH-Acked-by: Petr Matousek <pmatouse@redhat.com>
RH-Acked-by: Prarit Bhargava <prarit@redhat.com>
RH-Acked-by: Don Zickus <dzickus@redhat.com>

Bugzilla: 971315
https://bugzilla.redhat.com/show_bug.cgi?id=971315

upstream status
===============
commit f1923820c447e986a9da0fc6bf60c1dccdf0408e
Author: Stephane Eranian <eranian@google.com>
Date:   Tue Apr 16 13:51:43 2013 +0200

description
===========
The valid mask for both offcore_response_0 and
offcore_response_1 was wrong for SNB/SNB-EP,
IVB/IVB-EP. It was possible to write to
reserved bit and cause a GP fault crashing
the kernel.

This patch fixes the problem by correctly marking the
reserved bits in the valid mask for all the processors
mentioned above.

A distinction between desktop and server parts is introduced
because bits 24-30 are only available on the server parts.

This version of the  patch is just a rebase to perf/urgent tree
and should apply to older kernels as well.

brew
====
https://brewweb.devel.redhat.com/taskinfo?taskID=5938846

tests
=====
booted

Cc: Prarit Bhargava <prarit@redhat.com>
Cc: Don Zickus <dzickus@redhat.com>
Signed-off-by: Jarod Wilson <jarod@redhat.com>