gfs2: Reserve journal space for quota change in do_grow
Patchwork-id: 61009
O-Subject: [RHEL6.5 PATCH][bz976823] GFS2: Reserve journal space for quota change in do_grow
Bugzilla: 976823
RH-Acked-by: Abhijith Das <adas@redhat.com>
RH-Acked-by: Steven Whitehouse <swhiteho@redhat.com>

Hi,

If a GFS2 file system is mounted with quotas and a file is grown
in such a way that its free blocks for the allocation are represented
in a secondary bitmap, GFS2 ran out of blocks in the transaction.
That resulted in "fatal: assertion "tr->tr_num_buf <= tr->tr_blocks".
This patch reserves extra blocks for the quota change so the
transaction has enough space.

rhbz#976823

This patch appears upstream at this location:
http://git.kernel.org/cgit/linux/kernel/git/steve/gfs2-3.0-nmw.git/commit/fs/gfs2/bmap.c?id=a01aedfe21637c965a7046271fedfdd681eba646

It was tested on system gfs-a16c-01.mpc.lab.eng.bos.redhat.com with
a scenario that faithfully recreated the problem prior to the patch.

Regards,

Bob Peterson
Red Hat File Systems

Signed-off-by: Bob Peterson <rpeterso@redhat.com>
Signed-off-by: Rafael Aquini <aquini@redhat.com>