autofs: remove autofs dentry mount check
Patchwork-id: 61050
O-Subject: [RHEL-6.5 PATCH] autofs - remove autofs dentry mount check
Bugzilla: 947275
RH-Acked-by: Jeff Moyer <jmoyer@redhat.com>

Bugzilla bug: https://bugzilla.redhat.com/show_bug.cgi?id=947275

Upstream commit: c42127ed5aeb3feb266e1f8a180d9e6cdae2da0f

When checking if an autofs mount point is busy it isn't sufficient to
only check if it's a mount point.

For example, if the mount of an offset mountpoint in a tree is denied
for this host by its export and the dentry becomes a process working
directory the check incorrectly returns the mount as not in use at expire.

This can happen since the default when mounting within a tree is nostrict,
which means ingnore mount fails on mounts within the tree and continue. The
nostrict option is meant to allow mounting in this case.

Signed-off-by: David Jeffery <djeffery@redhat.com>
Signed-off-by: Ian Kent <ikent@redhat.com>
Signed-off-by: Rafael Aquini <aquini@redhat.com>