kernel: add support for physical memory > 4TB
Patchwork-id: 61092
O-Subject: [RHEL6.5 PATCH v3] [s390] kernel: add support for physical memory > 4TB
Bugzilla: 913220
RH-Acked-by: Prarit Bhargava <prarit@redhat.com>
RH-Acked-by: Steve Best <sbest@redhat.com>
RH-Acked-by: Don Zickus <dzickus@redhat.com>

Description
-----------
commit 14045ebf1e1156d966a796cacad91028e01797e5
Author: Martin Schwidefsky <schwidefsky@de.ibm.com>
Commit: Martin Schwidefsky <schwidefsky@de.ibm.com>

    [S390] add support for physical memory > 4TB

    The kernel address space of a 64 bit kernel currently uses a three level
    page table and the vmemmap array has a fixed address and a fixed maximum
    size. A three level page table is good enough for systems with less than
    3.8TB of memory, for bigger systems four page table levels need to be
    used. Each page table level costs a bit of performance, use 3 levels for
    normal systems and 4 levels only for the really big systems.
    To avoid bloating sparse.o too much set MAX_PHYSMEM_BITS to 46 for a
    maximum of 64TB of memory.

    Signed-off-by: Martin Schwidefsky <schwidefsky@de.ibm.com>

Dependencies
------------
This commit requires an update in the kexec-tools package (for the
makedumpfile tool) to deal with the increased address space.  This
dependency is tracked with bugzilla 825476 - "Scaling of dumps to
support 16S/12TB system".

Bugzilla
--------
BZ 913220
https://bugzilla.redhat.com/show_bug.cgi?id=913220

Upstream status of the patch
----------------------------
The patch is upstream as of kernel version 3.3
http://git.kernel.org/?p=linux/kernel/git/torvalds/linux-2.6.git;a=commitdiff;h=14045ebf1e1156d966a796cacad91028e01797e5

kABI
----
The increased MAX_PHYSMEM_BITS creates a kABI breakage in the mem_section
symbol in mm/sparse.c.  To solve this kABI issue, exlucde the change from
the kABI check.  As the MAX_PHYSMEM_BITS are increased, the size of the
mem_section increases. However, this should not be an issue for third party
modules (if they are any) for Linux on System z.

Test status
-----------
The patch has been tested and fixes the problem.
The fix has been verified by the IBM test department.

Signed-off-by: Rafael Aquini <aquini@redhat.com>