xfs: check on-disk, not incore, btree root size in dfrag.c
Patchwork-id: 61392
O-Subject: [RHEL6.5 PATCH 1/2] xfs: check on-disk (not incore) btree root size in dfrag.c
Bugzilla: 886566
RH-Acked-by: Carlos Maiolino <cmaiolino@redhat.com>
RH-Acked-by: Dave Chinner <dchinner@redhat.com>

For
Bug 886566 - xfs_swap_extents: inode <snip> format is incompatible for exchanging

Upstream patch currently in the oss.sgi.com xfs git tree:

>From 427d9fe233ccec3a293be4bcf07f9bac12463a99 Mon Sep 17 00:00:00 2001
From: Eric Sandeen <sandeen@redhat.com>
Date: Tue, 27 Mar 2012 15:40:26 -0500
Subject: [PATCH] xfs: check on-disk (not incore) btree root size in dfrag.c

xfs_swap_extents_check_format() contains checks to make sure that
original and the temporary files during defrag are compatible;
Gabriel VLASIU ran into a case where xfs_fsr returned EINVAL
because the tests found the btree root to be of size 120,
while the fork offset was only 104; IOW, they overlapped.

However, this is just due to an error in the
xfs_swap_extents_check_format() tests, because it is checking
the in-memory btree root size against the on-disk fork offset.
We should be checking the on-disk sizes in both cases.

This patch adds a new macro to calculate this size, and uses
it in the tests.

With this change, the filesystem image provided by Gabriel
allows for proper file degragmentation.

Reported-by: Gabriel VLASIU <gabriel@vlasiu.net>
Signed-off-by: Eric Sandeen <sandeen@redhat.com>
Reviewed-by: Dave Chinner <dchinner@redhat.com>
Signed-off-by: Ben Myers <bpm@sgi.com>
Signed-off-by: Rafael Aquini <aquini@redhat.com>