pcie: Report success only when every device has AER-aware driver
Patchwork-id: 61445
O-Subject: [RHEL6.5 PATCH 20/31] PCI/AER: Report success only when every device has AER-aware driver
Bugzilla: 911740
RH-Acked-by: Dean Nelson <dnelson@redhat.com>
RH-Acked-by: Prarit Bhargava <prarit@redhat.com>
RH-Acked-by: Don Dutile <ddutile@redhat.com>
RH-Acked-by: Tony Camuso <tcamuso@redhat.com>

Bugzilla: 911740
https://bugzilla.redhat.com/show_bug.cgi?id=911740
Upstream Status: 918b4053184c0ca22236e70e299c5343eea35304,
		 3c282db1656733c1267e6af4c22179bb04faf2ef

    There was an issue bringing this patch in upstream as it caused
    conflicts with ./drivers/pci/pcie/aer/aerdrv_core.c as can be seen by
    the follow-on (upstream) commit - 3c282db1656.  The follow-on patch is
    interesting in that it has much of the same content as this patch thus
    I didn't know what to do.  Normally I would have added an additional
    patch in this series for 3c282db1656 but that just didn't seem to make
    sense due to the duplication of most of the content.  I ended up just
    combining the result of applying both 918b4053184 (this patch) and
    3c282db1656.  I can change that if anyone objects.

    The differences related to:
      device_lock(&dev->dev);   ==> down(&dev->dev.sem);
      device_unlock(&dev->dev); ==> up(&dev->dev.sem);
    were explained in the previous patch of this series.

    NOTE: This series depends upon the following previous series -
      RHBZ736984

commit 918b4053184c0ca22236e70e299c5343eea35304
Author: Vijay Mohan Pandarathil <vijaymohan.pandarathil@hp.com>
Date:   Sat Nov 17 11:47:18 2012 +0000

PCI/AER: Report success only when every device has AER-aware driver

When an error is detected on a PCIe device which does not have an
AER-aware driver, prevent AER infrastructure from reporting
successful error recovery.

This is because the report_error_detected() function that gets
called in the first phase of recovery process allows forward
progress even when the driver for the device does not have AER
capabilities. It seems that all callbacks (in pci_error_handlers
structure) registered by drivers that gets called during error
recovery are not mandatory. So the intention of the infrastructure
design seems to be to allow forward progress even when a specific
callback has not been registered by a driver. However, if error
handler structure itself has not been registered, it doesn't make
sense to allow forward progress.

As a result of the current design, in the case of a single device
having an AER-unaware driver or in the case of any function in a
multi-function card having an AER-unaware driver, a successful
recovery is reported.

Typical scenario this happens is when a PCI device is detached
from a KVM host and the pci-stub driver on the host claims the
device. The pci-stub driver does not have error handling capabilities
but the AER infrastructure still reports that the device recovered
successfully.

The changes proposed here leaves the device(s)in an unrecovered state
if the driver for the device or for any device in the subtree
does not have error handler structure registered. This reflects
the true state of the device and prevents any partial recovery (or no
recovery at all) reported as successful.

[bhelgaas: changelog]
Signed-off-by: Vijay Mohan Pandarathil <vijaymohan.pandarathil@hp.com>
Signed-off-by: Bjorn Helgaas <bhelgaas@google.com>
Reviewed-by: Linas Vepstas <linasvepstas@gmail.com>
Reviewed-by: Myron Stowe <myron.stowe@redhat.com>

commit 3c282db1656733c1267e6af4c22179bb04faf2ef
Merge: d3fe398 918b405
Author: Bjorn Helgaas <bhelgaas@google.com>
Date:   Wed Nov 28 11:39:19 2012 -0700

    Merge branch 'pci/misc' into next

    * pci/misc:
      PCI/AER: Report success only when every device has AER-aware driver

    Conflicts:
        drivers/pci/pcie/aer/aerdrv_core.c