sfc: Properly sync RX DMA buffer when it is not the last in the page
Patchwork-id: 61730
O-Subject: [RHEL6.5 PATCH BZ 953929 53/95] sfc: Properly sync RX DMA buffer when it is not the last in the page
Bugzilla: 953929
RH-Acked-by: Stefan Assmann <sassmann@redhat.com>
RH-Acked-by: Tony Camuso <tcamuso@redhat.com>
RH-Acked-by: John Linville <linville@redhat.com>

From: Ben Hutchings <bhutchings@solarflare.com>

We may currently allocate two RX DMA buffers to a page, and only unmap
the page when the second is completed.  We do not sync the first RX
buffer to be completed; this can result in packet loss or corruption
if the last RX buffer completed in a NAPI poll is the first in a page
and is not DMA-coherent.  (In the middle of a NAPI poll, we will
handle the following RX completion and unmap the page *before* looking
at the content of the first buffer.)

Signed-off-by: Ben Hutchings <bhutchings@solarflare.com>
(cherry picked from commit 3a68f19d7afb80f548d016effbc6ed52643a8085)

Signed-off-by: Rafael Aquini <aquini@redhat.com>