make sure manual changes to recovery checkpoint are saved
Patchwork-id: 61863
O-Subject: [RHEL6.5 PATCH 040/105] md: make sure manual changes to recovery checkpoint are saved.
Bugzilla: 837097 869622
RH-Acked-by: Prarit Bhargava <prarit@redhat.com>
RH-Acked-by: Ivan Vecera <ivecera@redhat.com>
RH-Acked-by: Jonathan E Brassow <jbrassow@redhat.com>
RH-Acked-by: Stefan Assmann <sassmann@redhat.com>
RH-Acked-by: Tony Camuso <tcamuso@redhat.com>

From: NeilBrown <neilb@suse.de>

If you make an array bigger but suppress resync of the new region with
  mdadm --grow /dev/mdX --size=max --assume-clean

then stop the array before anything is written to it, the effect of
the "--assume-clean" is lost and the array will resync the new space
when restarted.
So ensure that we update the metadata in the case.

Reported-by: Sebastian Riemer <sebastian.riemer@profitbricks.com>
Signed-off-by: NeilBrown <neilb@suse.de>
(cherry picked from commit db07d85ef6395d4768c6cdceeb2ea4878bd2a3f4)

Signed-off-by: Jes Sorensen <Jes.Sorensen@redhat.com>
Signed-off-by: Rafael Aquini <aquini@redhat.com>