raid10: close race that lose writes lost when replacement completes
Patchwork-id: 61920
O-Subject: [RHEL6.5 PATCH 052/105] md/raid10: close race that lose writes lost when replacement completes.
Bugzilla: 837097 869622
RH-Acked-by: Prarit Bhargava <prarit@redhat.com>
RH-Acked-by: Ivan Vecera <ivecera@redhat.com>
RH-Acked-by: Jonathan E Brassow <jbrassow@redhat.com>
RH-Acked-by: Stefan Assmann <sassmann@redhat.com>
RH-Acked-by: Tony Camuso <tcamuso@redhat.com>

From: NeilBrown <neilb@suse.de>

When a replacement operation completes there is a small window
when the original device is marked 'faulty' and the replacement
still looks like a replacement.  The faulty should be removed and
the replacement moved in place very quickly, bit it isn't instant.

So the code write out to the array must handle the possibility that
the only working device for some slot in the replacement - but it
doesn't.  If the primary device is faulty it just gives up.  This
can lead to corruption.

So make the code more robust: if either  the primary or the
replacement is present and working, write to them.  Only when
neither are present do we give up.

This bug has been present since replacement was introduced in
3.3, so it is suitable for any -stable kernel since then.

Reported-by: "George Spelvin" <linux@horizon.com>
Cc: stable@vger.kernel.org
Signed-off-by: NeilBrown <neilb@suse.de>
(cherry picked from commit e7c0c3fa29280d62aa5e11101a674bb3064bd791)

Signed-off-by: Jes Sorensen <Jes.Sorensen@redhat.com>

Conflicts:
	drivers/md/raid10.c

Again caused by differences in plugging API

Signed-off-by: Rafael Aquini <aquini@redhat.com>