Remember the last sync operation that was performed
Patchwork-id: 61928
O-Subject: [RHEL6.5 PATCH 102/105] MD: Remember the last sync operation that was performed
Bugzilla: 837097 869622
RH-Acked-by: Prarit Bhargava <prarit@redhat.com>
RH-Acked-by: Ivan Vecera <ivecera@redhat.com>
RH-Acked-by: Jonathan E Brassow <jbrassow@redhat.com>
RH-Acked-by: Stefan Assmann <sassmann@redhat.com>
RH-Acked-by: Tony Camuso <tcamuso@redhat.com>

From: Jonathan Brassow <jbrassow@redhat.com>

MD:  Remember the last sync operation that was performed

This patch adds a field to the mddev structure to track the last
sync operation that was performed.  This is especially useful when
it comes to what is recorded in mismatch_cnt in sysfs.  If the
last operation was "data-check", then it reports the number of
descrepancies found by the user-initiated check.  If it was a
"repair" operation, then it is reporting the number of
descrepancies repaired.  etc.

Signed-off-by: Jonathan Brassow <jbrassow@redhat.com>
Signed-off-by: NeilBrown <neilb@suse.de>
(cherry picked from commit c4a39551451666229b4ea5e8aae8ca0131d00665)

Signed-off-by: Jes Sorensen <Jes.Sorensen@redhat.com>

Conflicts:
	Documentation/device-mapper/dm-raid.txt
	drivers/md/dm-raid.c
	drivers/md/md.c

Conflicts caused by lack of new plugging API in RHEL6. In addition
left out changes to dm-raid.c

Signed-off-by: Rafael Aquini <aquini@redhat.com>